requirements.txt
streamlit==1.39.0
python-docx==1.1.2
Pillow==10.4.0
